CIF-1444: Move '<AccountTrigger>' and '<AccountDropdown>' components to the folder of 'AccountContainer', update the import paths in '<AccountContainer>', this change allows us to group the account dropdown related components in same place (#303)
