+++ react-components/src/components/AccountContainer/__test__/accountTrigger.test.js
