+++ react-components/src/components/AccountContainer/accountTrigger.js
